refactor : 랜덤넘버생성부분을 전략패턴으로 변경 - 전략패턴으로 변경함으로서 해당부분이 테스트케이스에 포함된 부분일 경우 테스트용 전략패턴을 적용하여 테스트가 가능해진다.
