+++ src/main/java/lotto/domain/NumberGeneratorStrategy.java
@@ -1,0 +1,7 @@
+package lotto.domain;
+
+import java.util.List;
+
+public interface NumberGeneratorStrategy {
+    public List<String> generator();
+}
