+++ src/main/java/lotto/domain/RandomNumberGeneratorStrategy.java
@@ -1,23 +1,28 @@
-package lotto.utils;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class RandomNumberGenerator {
+public class RandomNumberGeneratorStrategy implements NumberGeneratorStrategy {
 
-    private RandomNumberGenerator() {
+    public RandomNumberGeneratorStrategy() {
     }
 
-    public static List<String> randomSixNumbers() {
+    @Override
+    public List<String> generator() {
+        return randomSixNumbers();
+    }
+
+    private List<String> randomSixNumbers() {
         List<String> availableNumbers = availableAllNumbers();
         Collections.shuffle(availableNumbers);
 
         return availableNumbers.subList(0, 6);
     }
 
-    private static List<String> availableAllNumbers() {
+    private List<String> availableAllNumbers() {
         return IntStream.range(1, 45)
                 .mapToObj(number -> String.valueOf(number))
                 .collect(Collectors.toList());
